requirements.txt
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.0.2
torch==1.11.0
tensorflow==2.8.0
tensorflow-federated==0.20.0
matplotlib==3.5.1
seaborn==0.11.2
flask==2.1.1
pycryptodome==3.14.1
python-dotenv==0.20.0
requests==2.27.1
networkx==2.7.1
PyYAML==6.0
pytest==7.1.1
jupyter==1.0.0
hyperledger-fabric-sdk==1.0.0
hfc==1.0.0
cryptography==36.0.2 